Allow templates specified from local filesystem
Fixes #47

Passing a template file (as opposed to a URL) to the CLI:

      ./heat -d create wordpress \
        --template-file=../templates/WordPress_Single_Instance.template

caused a 'URI too long' error. The reason was that we were passing the entire
template contents to the Heat API server inside the URL params.

This makes sure that templates are passed around inside HTTP body rather than
the URL.