Add 32-bit tdls for Fedora 16,17 jeos (gold and cfntools)
fixes #26

Note that the templates have been adjusted and the naming of the 32-bit
versions to i386 instead of i686. The install really is i686, but making
the arch match what is listed in the install ISO seems less confusing.

Signed-off-by: Jeff Peeler <jpeeler@redhat.com>

Although the 32-bit version of Fedora (and others) ISOs are labeled with i386,