Remove hard-coded paths from DB setup script
Since versions of heat can and will change, we do not want want them
hard-coded in the DB setup script. Instead, use the Python import logic to
determine the location of the module.

This will use the local copy of heat (which is usually what you want) if
run from the top level of the repository. Otherwise it will use the
installed version, as determined by Python.

Signed-off-by: Zane Bitter <zbitter@redhat.com>