Connect cfn utils to the remote server
The cfn_helper's `Metadata` class can now connect to the metadata server
prodived it's URL is known and available.

Engine will pass metadata server's URL via the User Data to the following file:

    /var/lib/cloud/data/cfn-metadata-server

If the file doesn't exist or the server is unreachable, cfntools will read
metadata from the `/var/lib/cloud/data/cfn-init-data` file as they did so far.

Note that Engine dosen't currently know the metadata server's URL so it's not
passing it to the instance yet.