Add a user creds database table and associate stacks with username.
This patch takes the credentials passed in from the context and allows
you to store them in the database in the 'user_creds' table for later
use with HA operations.  It also adds a 'username' to the stack table
for direct comparison and user validation to support per-user stacks.

Thanks to Angus for fixing the tests for me :)

Signed-off-by: Ian Main <imain@redhat.com>