Remove service name from nova() def
The service_name field is not unique between distros or versions of
openstack, but is used to select endpoints for communication with nova
and volume services.  The nova client will sort out the correct operation
based upon the service_type field.

The service_name field should probably just be removed from the API
definition upstream.

Tested on F16 & Devstack U12.

Change-Id: I36409dba9d9ec2b453a027fc1e2e78f7c8ace2a2
Signed-off-by: Steven Dake <sdake@redhat.com>