Implements a client side engine RPC API.
Patch 1/2, this implements github heat issue 181.

The topic string for the rpc call is built from the cfg.CONF parameters 'host'
and 'engine_topic' defined in heat.common.config.engine_opts.

delete_stack seems to be the only method which returns nothing, so it can be
invoked as cast or call, with cast being the default.

The tests exercise each api method with the same approach found in
nova/nova/tests/compute/test_rpcapi.py

Change-Id: Ia20850821083d1236ad628172db00d01f073313c