Make physical resource names unique based upon stack name
Some templates use the same resource name when registering with
nova.  Nova will not allow mulitple resource names to register
with the same name in all cases, and in those it does, it often
becomes confused (in the case of EIP).  This patch creates a
dot notation

For the LB test, an example of the resources would be:
teststack.LoadBalancer.LB_instance
teststack.DatabaseServer.MySqlDatabaseServer
teststack.WikiServerOne

Fixes Issue #160
Fixes Issue #136

Thanks to asalkeld@redhat.com for helping sort out the unit
test cases.

Change-Id: I14f4551feb41c7f98caa30dd98f997da6d00c467
Signed-off-by: Steven Dake <sdake@redhat.com>