Change rpc cleanup to occur before killing engine thread
This bug only manifested itself after a heat request was made to the
engine. Once a request was made a Qpid connection handler continued
to wait in a select call, presumably for a subsequent request. Closing
the connection immediately after detecting a keyboard interrupt and then
terminating the thread yields a clean shutdown with no traceback.

Closes #176

Change-Id: Ifdb82d33a6c52a004bbd6b08c564264f4140e800
Signed-off-by: Jeff Peeler <jpeeler@redhat.com>