Avoid create_stack RPC call timing out on precise
The create_stack RPC call times out on Ubuntu as novaclient
blocks expecting stdin when prompting for a keyring password:

    Please set a password for your new keyring

This probably does not occur on Fedora due to the version or
presence of the python-keyring package differing.

The problem can be avoided by ensuring the no_cache flag is
set on the novaclient.

Change-Id: I3ac9e74571ff837cd07fc3f36ecf706180f4dc46