heat workaround for HA/Autoscaling regression
Interim workaround for HA/Autoscaling regression, caused by change of
scope for stack_get_by_name to be per-tenant, meaning we need to do a
brute-force lookup to find the stack entry when using the stored admin
context.

A better, more comprehensive (but more intrusive) fix is in-progress
but this at least returns us to the same functionality we had before

Note this means that all stacks using WatchRule still need to have names
unique accross all tenants.

bug 1078779

Change-Id: Iad1830d38262d1afb63cee16a3e366d9fd09acb4
Signed-off-by: Steven Hardy <shardy@redhat.com>