Handle duplicate physical resources IDs
Because the unit tests create multiple stacks in different tenants with the
same physical resource ID, the database lookup code must change to
accomodate this.

This enables us to add a unit test for the describe_stack_resources RPC
API call for the case where we pass a physical resource ID to lookup.

Change-Id: Id37b15bee595ba10f207235a3ff59526e4423be6
Signed-off-by: Zane Bitter <zbitter@redhat.com>