Move abstract nested stack class to separate file
The class heat.engine.resources.stack.Stack is intended to be an abstract
class which resources that are implemented as nested stacks may inherit
from. Rename it to StackResource and move it to the
heat.engine.stack_resource module, so that other resource modules
(including plugins) may include it without having to also import the
NestedStack class (which implements the AWS::CloudFormation::Stack resource
type).

Change-Id: Ie7c05ec64e1c9632c7a24d96b5d790c9e50b691f
Signed-off-by: Zane Bitter <zbitter@redhat.com>