A native Swift container resource type.
Implements blueprint swift-resource-type

The properties schema maps to the swift headers
that are set during container creation.

The schema allows any arbitrary X-Container-Meta-* header to
be set in the template, which will be useful for users who
store arbitrary header keys.

This will be used during the presentation demo at LCA
next week.

Change-Id: I35e594cbe3cc251d5b48b41ea78ed2edda8aa807