heat engine : Compare runtime resolved resource snippets on update
We need to compare the runtime resolved resource snippet, since that
is what we get back from self.parsed_template.  If we don't do this,
then we get a false positive when a Ref to a parameter which gets
updated (e.g AWS::StackId) is used in the resource properties.

ref bug 1131666

Change-Id: Ib488c43b9eca998a7a82b7571097f5bb69ef946c