Fix flaky unit test
This was failing intermittently during the gate tests because there was no
dependency in the template for the Instance on the WaitConditionHandle, nor
for the WaitCondition on the Instance, so the order in which resources
would be created was somewhat random.

This patch fixes the test template, and tests updating the metadata should
occur *during* creation of the stack, as well after, to replicate a real
wait condition (which won't finish creation until the metadata update
occurs).

Change-Id: I26b5590a9ee90d205bd9f2e31bcb74e68065ac25
Signed-off-by: Zane Bitter <zbitter@redhat.com>