Avoid manipulating DB directly in unit test
Use the parser.Stack class for inserting data into the database in the
watch rule unit tests. If there is direct database manipulation in the unit
tests, it risks falling out of step with the implementation in the code (in
this case, by not supplying the disable_rollback value).

(Required to squash DB migrations, bug 1072949)

Change-Id: I66e5a0093d20222c876f52e626043704777dcb6a