validate_template returns whole Parameters snippet
Currently the engine validate_template returns cfn formatted
Parameters from the template, which isn't actually very useful
to build a full UI from a validate call.

This change returns the original Parameters template snippet
and changes the cfn api to reformat to the CFN API legacy format.

The Parameters section of the RPC and the REST APIs now return
a dictionary containing the parameters as specified in the template.
Previously the Parameters section contained a list in the CFN API
format.

Pseudo parameters are filtered out of the validate results.

Change-Id: Iea53dc847ff13f5f479ec3a66bcf141c4ccaeb25
Fixes: bug #1157537