Allow nova operations with only auth_token
This uses the same technique as Horizon for auth_token operations.

Specifically, the nova endpoint is found directly from keystone
and the auth_token and management_url are set on novaclient
after it has been created.

Note that the explicit call to authenticate() is not required
and actually breaks the auth_token case.

Tested by launching a stack with and without the --token-only flag

Also tested using Horizon, which is auth_token only

Part of blueprint auth-token-only

Change-Id: Ic11a838b96a4b5130cb2ac9d685e8a97c9d0a0ad