Log tracepath for stack validate exceptions.
Calling validate on each resource exercises a lot of code, and
will at times cause non-validation related exceptions.

This change just logs the tracepath for any exception that happens
during validate.

This may be a contender for backporting to milestone-proposed,
since diagnosing some problems on the grizzly branch will be
difficult without it

Change-Id: Ib010a8ec826552e2817da3a22c633dcac6c3ffc6
Fixes: bug #1163586