Subclass keystone middleware to set headers
Replace old forked auth_token with a subclass
of keystoneclient middleware.

The advantages of moving to keystoneclient middleware:
- can use v3 (or v2) keystone api
- PKI tokens
- token revocation

The subclass sets the following headers to be consumed by our
request context filter:
- X-Admin-User
- X-Admin-Pass
- X-Admin-Tenant-Name
- X-Auth-Url

The need to override _build_user_headers should be a short-term thing,
X-Admin-* isn't actually used currently, and there are a few options
that need to be discussed for getting X-Auth-Url to the engine.

Change-Id: Iacc5046fbf559724a4ae0bd6091d662e23d65544
Blueprint: keystone-middleware