Fix AccessKey deletion with bad credentials
If an AccessKey could not be created, it also could not be deleted. This
change allows deletion of the resource to proceed even if the user does not
exist.

Fixes bug 1155824

Also fixes a bug that the actual key would never be deleted, because the
resource_id is set to None beforehand.

Change-Id: I34886410e3e25ad62e4eba537063f922b011e25a