Simplify rpc client calls
We always pass host=None to _engine_topic() which means that we always
construct a topic using CONF.engine_topic and CONF.host.

And since RpcProxy methods falls back to the topic supplied to its
constructor, we never need to explicitly pass it.

We can safely remove the _engine_topic() method and the passing of an
explicit topic to RpcProxy methods and replace it with constructing the
correct topic at construction time, without causing any functional
changes.

While we're at it, replace the last use of the variable name FLAGS with
CONF.

Change-Id: Iefdc6b78a068975ac45b10d6b53210cb283d2ef8