Mark resources as failed when creation aborted
Previously resources would be left in the CREATE_IN_PROGRESS state in the
event of a timeout or the thread being cancelled (due to a delete being
requested while the stack was still being created). We should instead put
these resources into the CREATE_FAILED state when this occurs, thus
creating an event with the reason.

Change-Id: I05956aa9ef26941b79cef9dbf0cb6a347047d199