Do not initialize anything during import phase
The structure of the program is compromised by doing too much logic in
the import phase. We can read the code more cleanly if we can see where
the intended initialization code is called. This may make tests slightly
more tedious to write, but that should be handled by writing generic
fixtures which can be reused for exactly this purpose.

Change-Id: I1d221e2f90d1f6e89717a37d2128caabd077d30a