Removing all prints and capturing logging
A new base class for tests is used to reduce repetitive steps used in
a majority of tests. This new base class also uses fixtures.FakeLogger
to suppress logging, though it will be shown on any failures to aid in
debugging the failure.

Prints that happen on normal operation of the test suite are all removed
as they are not needed and only clutter the output.

Change-Id: I0365283ce415c5390fd68bdf1f0b3c8038b1b9af