Send RPCs to 'engine' topic not 'engine.$host'
Fixes bug #1173573

We currently send engine RPCs to the 'engine.$host' topic means that
it is impossible to run heat-api and heat-engine on different
machines because you can't configure heat-api with the hostname of
heat-engine.

Fix this simply by sending RPCs to 'engine' which heat-engine. We have
always listened on this topic, so there is no backwards compat
concerns.

This is somewhat related to the multiple-engines blueprint, but
doesn't really help or hinder it.

Change-Id: Ibd4dc948ace4476bfa8e7fdc86e948a622416e58