Implement the SubnetId property in the Instance resource
If the SubnetId parameter is given in the template, the handle_create method
in the Instance resource verifies the subnet in quantum and configures the
NetworkInterfaces for the subnet. If no NetworkInterface is specified, a port
is created for the subnet and the nic is configured to use that port.

Change-Id: I5a05cbe0798e6b181bf9dd3b547cda525e18480f
Fixes: bug #1163952