Rename handle_snapshot and pass state
Rename handle_snapshot() to handle_snapshot_delete() to better reflect what
it needs to do (both take a snapshot of the resource and then delete the
resource). Also, pass the state of the resource prior to delete being
invoked, since the resource may want to use different error handling in the
snapshot phase depending on e.g. whether the resource was created
successfully or not.

Change-Id: I1e5aaab4f13e42333c1f2f39bfeb0ebf2ac89ed8