Rework associations from vpc to quantum resources.
Metadata cannot be used to store the relationships between VPC
resources and the underlying quantum resources, since this prevents
VPC or Subnet refs from being passed in as parameters.

This rework results in no state being stored in metadata. Route table
relationships are inferred from the template. An assumption is made that
the quantum router associatiated with the VPC will have the same name
as the quantum net. The easiest way of ensuring that is to creat the VPC
in a different Heat stack.

Fixes bug: #1165050
Fixes bug: #1165056
Fixes bug: #1166779
Fixes bug: #1166787

Change-Id: I6fb087bfc63ed13703089a16ed57858e6911b7d8