Add dependency between RouterGateway and RouterInterface
When a template has a RouterGateway and a RouterInterface with the
same router_id, there is an implicit dependency on delete where
the gateway must be deleted before the interface.

This is implemented by overiding add_dependencies in RouterGateway
and looking in the stack for RouterInterface resource to depend on.

Fixed bug: #1182266

Change-Id: I90ebe987c1d6e81f8936ceb81b81376c7e0d497a