Add dependency between FloatingIP and RouterGateway
When a template has a FloatingIP and a RouterGateway on the same
external network, there is an implicit dependency on create where
the gateway must be created before the floating ip.

This is implemented by overiding add_dependencies in FloatingIP
and looking in the stack for RouterGateway resource to depend on.

Fixed bug: #1182266

Change-Id: I40f281e5bf2a27280d328fe11d32e6404f412a79