Support cloud-specific suffixes to server status.
Some clouds, *cough*HP*cough* append extra data to their
status strings. Such as: BUILD(scheduling) or BUILD(networking).
We don't really want to be on the hook for keeping a list of those,
but if we just grab the bit in front of the parenthetical, we should
be able to tell state sensibly, even when people are being weird.

Change-Id: Idaf079d6090db50c9041f398a257b427d63ff48d