Emit alarm actions if a rule remains in ALARM state.
It looks like the original intent of this edge-triggering was to
prevent actions being emitted too frequently, causing overshoot of
scaling resources.

However this causes an issue when one scaling resource change is not
enough to bring the alarm state under the threshold.

Proper use of Period, EvaluationPeriods and Threshold should be sufficient
to prevent scaling resource overshoot.

Fixes bug: #1193970

Change-Id: I5b38eb82cd1ad568fe4af77999e9d52f9eb99328