autoscaling : Fix issue when scaling to zero instances
As noted in https://review.openstack.org/#/c/34564, we lack
coverage of the scenario when MinSize is 0 and we scale down
to zero instances then try to do a delete.  We do the wrong
thing here, by creating a GroupedInstance object with name '',
and this would also result in an incorrect update to the
LoadBalancer.  Fix is to reset the resource_id to None.

Fixes bug #1195364
Change-Id: Ie3a52badf35f7e660de9a8f38635d2a09463123a